+++ app/src/main/java/com/eatssu/android/data/dto/request/ChangeNicknameRequestDto.kt
@@ -1,4 +1,4 @@
-package com.eatssu.android.data.model.request
+package com.eatssu.android.data.dto.request
 
 data class ChangeNicknameRequestDto(
     val nickname: String
