+++ app/src/main/java/com/eatssu/android/data/dto/request/ModifyMenuRequestDto.kt
@@ -1,4 +1,4 @@
-package com.eatssu.android.data.model.request
+package com.eatssu.android.data.dto.request
 
 data class ModifyMenuRequestDto(
     val addTodayMenuList: List<AddTodayMenu>
