+++ app/src/main/java/com/eatssu/android/data/dto/request/WriteReviewDetailRequestDto.kt
@@ -1,4 +1,4 @@
-package com.eatssu.android.data.model.request
+package com.eatssu.android.data.dto.request
 
 import com.eatssu.android.ui.review.write.Review
