+++ app/src/main/java/com/eatssu/android/data/dto/response/TokenResponseDto.kt
@@ -1,4 +1,4 @@
-package com.eatssu.android.data.model.response
+package com.eatssu.android.data.dto.response
 
 import com.google.gson.annotations.SerializedName
