+++ app/src/main/java/com/eatssu/android/data/model/CalendarData.kt
@@ -1,4 +1,4 @@
-package com.eatssu.android.data.entity
+package com.eatssu.android.data.model
 
 data class CalendarData(
     var cl_date: String = "", // 날짜
