+++ app/src/main/java/com/eatssu/android/data/model/Menu.kt
@@ -1,4 +1,4 @@
-package com.eatssu.android.data.entity
+package com.eatssu.android.data.model
 
 data class Menu(
     val id: Long,
