+++ app/src/main/java/com/eatssu/android/data/service/MyPageService.kt
@@ -1,11 +1,9 @@
 package com.eatssu.android.data.service
 
-import com.eatssu.android.data.model.response.GetMyInfoResponseDto
-import com.eatssu.android.data.model.response.GetMyReviewResponseDto
+import com.eatssu.android.data.dto.response.GetMyInfoResponseDto
+import com.eatssu.android.data.dto.response.GetMyReviewResponseDto
 import retrofit2.Call
 import retrofit2.http.GET
-import retrofit2.http.POST
-import retrofit2.http.Query
 
 interface MyPageService {
     @GET("mypage/myreview") //내가 쓴 리뷰 모아보기
