+++ app/src/main/java/com/eatssu/android/ui/common/VersionViewModel.kt
@@ -2,7 +2,7 @@ package com.eatssu.android.ui.common
 
 import androidx.lifecycle.ViewModel
 import com.eatssu.android.BuildConfig.VERSION_NAME
-import com.eatssu.android.data.entity.AndroidMessage
+import com.eatssu.android.data.model.AndroidMessage
 import com.eatssu.android.data.repository.FirebaseRemoteConfigRepository
 
 class VersionViewModel(private val repository: FirebaseRemoteConfigRepository) : ViewModel() {
