+++ app/src/main/java/com/eatssu/android/ui/login/LoginViewModel.kt
@@ -4,9 +4,9 @@ import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
-import com.eatssu.android.data.model.request.LoginWithKakaoRequestDto
+import com.eatssu.android.data.dto.request.LoginWithKakaoRequestDto
 import com.eatssu.android.base.BaseResponse
-import com.eatssu.android.data.model.response.TokenResponseDto
+import com.eatssu.android.data.dto.response.TokenResponseDto
 import com.eatssu.android.data.service.OauthService
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.launch
