+++ app/src/main/java/com/eatssu/android/ui/main/menu/MenuAdapter.kt
@@ -6,7 +6,7 @@ import android.view.ViewGroup
 import androidx.core.content.ContextCompat
 import androidx.recyclerview.widget.LinearLayoutManager
 import androidx.recyclerview.widget.RecyclerView
-import com.eatssu.android.data.entity.Section
+import com.eatssu.android.data.model.Section
 import com.eatssu.android.databinding.ItemSectionBinding
 import com.eatssu.android.ui.info.InfoActivity
