+++ app/src/main/java/com/eatssu/android/ui/main/menu/MenuFragment.kt
@@ -12,14 +12,14 @@ import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModelProvider
 import androidx.recyclerview.widget.LinearLayoutManager
 import com.eatssu.android.App
-import com.eatssu.android.data.entity.Menu
-import com.eatssu.android.data.entity.Section
+import com.eatssu.android.data.model.Menu
+import com.eatssu.android.data.model.Section
 import com.eatssu.android.data.enums.MenuType
 import com.eatssu.android.data.enums.Restaurant
 import com.eatssu.android.data.enums.Time
-import com.eatssu.android.data.model.response.GetFixedMenuResponseDto
-import com.eatssu.android.data.model.response.GetTodayMeal
-import com.eatssu.android.data.model.response.GetTodayMealResponseDto
+import com.eatssu.android.data.dto.response.GetFixedMenuResponseDto
+import com.eatssu.android.data.dto.response.GetTodayMeal
+import com.eatssu.android.data.dto.response.GetTodayMealResponseDto
 import com.eatssu.android.data.service.MenuService
 import com.eatssu.android.databinding.FragmentMenuBinding
 import com.eatssu.android.ui.main.calendar.CalendarViewModel
