+++ app/src/main/java/com/eatssu/android/ui/main/menu/MenuSubAdapter.kt
@@ -6,7 +6,7 @@ import android.view.LayoutInflater
 import android.view.ViewGroup
 import androidx.core.content.ContextCompat
 import androidx.recyclerview.widget.RecyclerView
-import com.eatssu.android.data.entity.Menu
+import com.eatssu.android.data.model.Menu
 import com.eatssu.android.data.enums.MenuType
 import com.eatssu.android.databinding.ItemMenuBinding
 import com.eatssu.android.ui.review.list.ReviewActivity
