+++ app/src/main/java/com/eatssu/android/ui/mypage/myreview/MyReviewAdapter.kt
@@ -8,7 +8,7 @@ import android.widget.ImageView
 import androidx.core.content.ContextCompat
 import androidx.recyclerview.widget.RecyclerView
 import com.bumptech.glide.Glide
-import com.eatssu.android.data.model.response.GetMyReviewResponseDto
+import com.eatssu.android.data.dto.response.GetMyReviewResponseDto
 import com.eatssu.android.databinding.ItemReviewBinding
 import com.eatssu.android.ui.review.etc.MyReviewDialogActivity
 import com.eatssu.android.util.MySharedPreferences
