+++ app/src/main/java/com/eatssu/android/ui/mypage/myreview/MyReviewListActivity.kt
@@ -5,7 +5,7 @@ import android.util.Log
 import android.view.View
 import androidx.recyclerview.widget.LinearLayoutManager
 import com.eatssu.android.base.BaseActivity
-import com.eatssu.android.data.model.response.GetMyReviewResponseDto
+import com.eatssu.android.data.dto.response.GetMyReviewResponseDto
 import com.eatssu.android.data.service.MyPageService
 import com.eatssu.android.databinding.ActivityMyReviewListBinding
 import com.eatssu.android.util.RetrofitImpl.retrofit
