+++ app/src/main/java/com/eatssu/android/ui/mypage/usernamechange/UserNameChangeViewModel.kt
@@ -4,7 +4,7 @@ import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
-import com.eatssu.android.data.model.request.ChangeNicknameRequestDto
+import com.eatssu.android.data.dto.request.ChangeNicknameRequestDto
 import com.eatssu.android.base.BaseResponse
 import com.eatssu.android.data.service.UserService
 import kotlinx.coroutines.Dispatchers
