+++ app/src/main/java/com/eatssu/android/ui/review/list/ReviewAdapter.kt
@@ -8,7 +8,7 @@ import android.widget.ImageView
 import androidx.core.content.ContextCompat
 import androidx.recyclerview.widget.RecyclerView
 import com.bumptech.glide.Glide
-import com.eatssu.android.data.model.response.GetReviewListResponse
+import com.eatssu.android.data.dto.response.GetReviewListResponse
 import com.eatssu.android.databinding.ItemOthersReviewBinding
 import com.eatssu.android.databinding.ItemReviewBinding
 import com.eatssu.android.ui.review.etc.MyReviewDialogActivity
